feat: Retry strategy for connection errors #34
