+++ src/InfluxDB2/DefaultApi.php
@@ -87,6 +87,7 @@ class DefaultApi
 
         } catch (RequestException $e) {
             throw new ApiException(
+                $e,
                 "[{$e->getCode()}] {$e->getMessage()}",
                 $e->getCode(),
                 $e->getResponse() ? $e->getResponse()->getHeaders() : null,
